GaurdRail_FastAPI/requirements.txt
fastapi
uvicorn
guardrails-ai
openai
pydantic
presidio-analyzer
presidio-anonymizer
spacy
python-dotenv
